refactor(execution): reuse shared Unity path helpers
Why:
- the project domain introduced duplicate Unity project-root resolution and path normalization logic
- keeping those copies in sync would make future path behavior changes easy to miss

What Changed:
- added a shared UnityProjectPathResolver for editor-side project root resolution
- updated project operations and plan-token environment to use the shared resolver
- replaced ad-hoc slash and platform separator conversions in the new project-domain paths with existing path helpers and aligned affected tests

Impact:
- Unity project root resolution now has one implementation in the editor assembly
- project-domain path normalization reuses the existing contracts helper instead of maintaining another copy
- the open PR is updated without changing the refresh/save behavior contract
