+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/UnityProjectPathResolver.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2dc490ed7828445a28f636a9cabb297a
